--- a/paddle/phi/kernels/gpu/expand_as_grad_kernel.cu
+++ b/paddle/phi/kernels/gpu/expand_as_grad_kernel.cu
@@ -15,8 +15,46 @@
 #include "paddle/phi/kernels/expand_as_grad_kernel.h"
 
 #include "paddle/phi/backends/gpu/gpu_context.h"
+#include "paddle/phi/core/dense_tensor.h"
 #include "paddle/phi/core/kernel_registry.h"
-#include "paddle/phi/kernels/impl/expand_as_grad_kernel_impl.h"
+#include "paddle/phi/kernels/funcs/reduce_function.h"
+
+#define MAX_RANK_SUPPORTED 6
+
+namespace phi {
+
+template <typename T, typename Context>
+void ExpandAsGradKernel(const Context& context,
+                        const DenseTensor& x,
+                        const DenseTensor& out_grad,
+                        const std::vector<int>& target_shape,
+                        DenseTensor* in_grad) {
+  auto in_dims = x.dims();
+  auto out_dims = out_grad.dims();
+  int in_rank = in_dims.size();
+  int out_rank = out_dims.size();
+
+  PADDLE_ENFORCE_LE(
+      out_rank,
+      MAX_RANK_SUPPORTED,
+      errors::InvalidArgument("The rank of the input 'Out@GRAD' for "
+                              "expand_as_v2_grad op must be less than or equal "
+                              "to %d, but the value received is %d.",
+                              MAX_RANK_SUPPORTED,
+                              out_rank));
+
+  context.template Alloc<T>(in_grad);
+  if (in_dims == out_dims) {
+    phi::Copy(context, out_grad, context.GetPlace(), false, in_grad);
+  } else {
+    std::vector<int> reduce_dims = funcs::GetReduceDim(in_dims, out_dims, -1);
+
+    funcs::ReduceKernel<T, T, kps::AddFunctor, kps::IdentityFunctor<T>>(
+        context, out_grad, in_grad, kps::IdentityFunctor<T>(), reduce_dims);
+  }
+}
+
+}  // namespace phi
 
 PD_REGISTER_KERNEL(expand_as_grad,
                    GPU,