@Timber-Ye 2023-04-11
> @Timber-Ye @BrianQian1999 经过黑客松组委会讨论：
> 
> 1. 辛苦将fp16注册的代码移除，移除后的PR review通过后即可合入
> 2. 如果有兴趣可以继续解fp16注册的问题，解决后额外再送一个小礼品
> 
> cc @JamesLim-sy @zhangbopd

Done.